+++ js/adapt-contrib-boxMenuView.js
@@ -1,0 +1,151 @@
+import Adapt from 'core/js/adapt';
+import MenuView from 'core/js/views/menuView';
+import BoxMenuItemView from './adapt-contrib-boxMenuItemView';
+import BoxMenuGroupView from './adapt-contrib-boxMenuGroupView';
+
+class BoxMenuView extends MenuView {
+
+  className() {
+    return 'boxmenu';
+  }
+
+  initialize() {
+    super.initialize();
+    this.setStyles();
+
+    this.listenTo(Adapt, 'device:changed', this.onDeviceResize);
+  }
+
+  onDeviceResize() {
+    this.setStyles();
+  }
+
+  addChildren() {
+    let nthChild = 0;
+    const models = this.model.getChildren().models;
+    const childViews = [];
+    models.forEach(model => {
+      if (!model.get('_isAvailable')) return;
+
+      nthChild++;
+      model.set('_nthChild', nthChild);
+
+      const ChildView = (model.get('_type') === 'menu' && model.get('_boxMenu') && model.get('_boxMenu')._renderAsGroup) ?
+        BoxMenuGroupView :
+        BoxMenuItemView;
+
+      const $parentContainer = this.$(this.constructor.childContainer);
+      const childView = new ChildView({ model });
+
+      childViews.push(childView);
+
+      $parentContainer.append(childView.$el);
+    });
+
+    this.setChildViews(childViews);
+  }
+
+  setStyles() {
+    this.setBackgroundImage();
+    this.setBackgroundStyles();
+    this.processHeader();
+  }
+
+  setBackgroundImage() {
+    const config = this.model.get('_boxMenu');
+    const backgroundImages = config && config._backgroundImage;
+
+    if (!backgroundImages) return;
+
+    const backgroundImage = backgroundImages[`_${Adapt.device.screenSize}`] ?? backgroundImages._small;
+
+    if (backgroundImage) {
+      this.$el
+        .addClass('has-bg-image')
+        .css('background-image', 'url(' + backgroundImage + ')');
+      return;
+    }
+    this.$el
+      .removeClass('has-bg-image')
+      .css('background-image', '');
+  }
+
+  setBackgroundStyles() {
+    const config = this.model.get('_boxMenu');
+    const styles = config && config._backgroundStyles;
+
+    if (!styles) return;
+
+    this.$el.css({
+      'background-repeat': styles._backgroundRepeat,
+      'background-size': styles._backgroundSize,
+      'background-position': styles._backgroundPosition
+    });
+  }
+
+  processHeader() {
+    const config = this.model.get('_boxMenu');
+    const header = config && config._menuHeader;
+
+    if (!header) return;
+
+    const $header = this.$('.menu__header');
+
+    this.setHeaderBackgroundImage(header, $header);
+    this.setHeaderBackgroundStyles(header, $header);
+    this.setHeaderMinimumHeight(header, $header);
+  }
+
+  setHeaderBackgroundImage(config, $header) {
+    const backgroundImages = config._backgroundImage;
+
+    if (!backgroundImages) return;
+
+    const backgroundImage = backgroundImages[`_${Adapt.device.screenSize}`] ?? backgroundImages._small;
+
+    if (backgroundImage) {
+      $header
+        .addClass('has-bg-image')
+        .css('background-image', 'url(' + backgroundImage + ')');
+      return;
+    }
+    $header
+      .removeClass('has-bg-image')
+      .css('background-image', '');
+  }
+
+  setHeaderBackgroundStyles(config, $header) {
+    const styles = config._backgroundStyles;
+
+    if (!styles) return;
+
+    $header.css({
+      'background-repeat': styles._backgroundRepeat,
+      'background-size': styles._backgroundSize,
+      'background-position': styles._backgroundPosition
+    });
+  }
+
+  setHeaderMinimumHeight(config, $header) {
+    const minimumHeights = config._minimumHeights;
+
+    if (!minimumHeights) return;
+
+    const minimumHeight = minimumHeights[`_${Adapt.device.screenSize}`] ?? minimumHeights._small;
+
+    if (minimumHeight) {
+      $header
+        .addClass('has-min-height')
+        .css('min-height', minimumHeight + 'px');
+      return;
+    }
+    $header
+      .removeClass('has-min-height')
+      .css('min-height', '');
+  }
+
+}
+
+BoxMenuView.template = 'boxMenu';
+
+export default BoxMenuView;
